refactor: supabase SSR로직을 위한 파일 생성(api/supabaseSSR/supabase.ts) 및 Profile.tsx supabase SSR 로직 분리
